Group operation builders into a single module
