+++ lib/meow/ops.ex
@@ -1,9 +1,89 @@
-defmodule Meow.Op.Multi do
+defmodule Meow.Ops do
   @moduledoc """
-  Core operations relevant to multi-population algorithms.
+  Core operations universal for all representation types.
   """
 
-  alias Meow.{Op, Pipeline, Population, Topology}
+  alias Meow.{Op, Pipeline, Population}
+
+  @doc """
+  Builds an operation that terminates the population
+  if the given number of generations is reached.
+  """
+  @doc type: :termination
+  @spec max_generations(non_neg_integer()) :: Op.t()
+  def max_generations(generations) do
+    %Op{
+      name: "Termination: max generations",
+      requires_fitness: false,
+      invalidates_fitness: false,
+      impl: fn population, _ctx ->
+        if population.generation >= generations do
+          %{population | terminated: true}
+        else
+          population
+        end
+      end
+    }
+  end
+
+  @doc """
+  Builds an operation that introduces branching into the pipeline.
+
+  The operation splits population into a number of populations,
+  each of which is then passed through a corresponding pipeline
+  from the given list. Finally all populations are joined back
+  into a single population with the given join function.
+  """
+  @doc type: :flow
+  @spec split_join(
+          (Population.t() -> list(Population.t())),
+          list(Pipeline.t()),
+          (list(Population.t()) -> Population.t())
+        ) :: Op.t()
+  def split_join(split_fun, pipelines, join_fun) do
+    requires_fitness = Enum.any?(pipelines, fn %{ops: [op | _]} -> op.requires_fitness end)
+
+    %Op{
+      name: "Flow: split join",
+      # This operation itself doesn't require fitness,
+      # but if any pipeline does, then we eagerly do so as well.
+      requires_fitness: requires_fitness,
+      # This operation itself doesn't invalidate fitness,
+      # it just joins results of the underlying pipelines.
+      invalidates_fitness: false,
+      impl: fn population, ctx ->
+        population
+        |> split_fun.()
+        |> Enum.zip(pipelines)
+        |> Enum.map(fn {population, pipeline} ->
+          Pipeline.apply(population, pipeline, ctx)
+        end)
+        |> join_fun.()
+      end
+    }
+  end
+
+  @doc """
+  Builds an operation that introduces conditional flow
+  into the pipeline.
+
+  The operation evaluates the given predicate function
+  and passes the population through either of the given
+  pipelines.
+  """
+  @doc type: :flow
+  @spec if((Population.t() -> boolean()), Pipeline.t(), Pipeline.t()) :: Op.t()
+  def if(pred_fun, on_true_pipeline, on_false_pipeline) do
+    %Op{
+      name: "Flow: if",
+      requires_fitness: false,
+      invalidates_fitness: false,
+      impl: fn population, ctx ->
+        pipeline = if(pred_fun.(population), do: on_true_pipeline, else: on_false_pipeline)
+        Pipeline.apply(population, pipeline, ctx)
+      end
+    }
+  end
 
   @doc """
   Builds an emigration operation.
@@ -23,6 +103,7 @@ defmodule Meow.Op.Multi do
       or a range, in which case the number is randomly
       drawn every time. Defaults to 1.
   """
+  @doc type: :multi
   @spec emigrate(Op.t(), Topology.topology_fun(), keyword()) :: Op.t()
   def emigrate(selection_op, topology_fun, opts \\ []) do
     interval = Keyword.get(opts, :interval, 1)
@@ -105,6 +186,7 @@ defmodule Meow.Op.Multi do
       indicates that the algorithm run into deadlock
       (populations waiting for each other). Defaults to `20_000`.
   """
+  @doc type: :multi
   @spec immigrate((non_neg_integer() -> Op.t()), keyword()) :: Op.t()
   def immigrate(size_to_selection_op, opts \\ []) do
     interval = Keyword.get(opts, :interval, 1)
